# ServerApp/API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/addExercise": {
      "post": {
        "tags": [
          "Exercise"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Exercise/getAllExercises": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetExerciseListItemQuery"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetExerciseListItemQuery"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetExerciseListItemQuery"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/getExercise/{id}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetExerciseQuery"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExerciseQuery"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExerciseQuery"
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/editExercise/{id}": {
      "put": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditExerciseCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditExerciseCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditExerciseCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Stats/addStats/exercise={exerciseId}": {
      "post": {
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseUnitCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseUnitCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseUnitCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Stats/getStats/exercise={exerciseId}": {
      "get": {
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetExerciseUnitListItemQuery"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetExerciseUnitListItemQuery"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetExerciseUnitListItemQuery"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Stats/getLastStats/exercise={exerciseId}": {
      "get": {
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetExerciseUnitItemQuery"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExerciseUnitItemQuery"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExerciseUnitItemQuery"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/register": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/User/changepassword": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateExerciseCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "extraSet": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CreateExerciseUnitCommand": {
        "type": "object",
        "properties": {
          "sessionDate": {
            "type": "string",
            "format": "date-time"
          },
          "exerciseSet": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseSet"
            },
            "nullable": true
          },
          "extraSet": {
            "$ref": "#/components/schemas/ExerciseSet"
          }
        },
        "additionalProperties": false
      },
      "EditExerciseCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "extraSet": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ExerciseSet": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "format": "float"
          },
          "reps": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetExerciseListItemQuery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetExerciseQuery": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "extraSet": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "GetExerciseUnitItemQuery": {
        "type": "object",
        "properties": {
          "sessionDate": {
            "type": "string",
            "format": "date-time"
          },
          "exerciseSet": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseSet"
            },
            "nullable": true
          },
          "extraSet": {
            "$ref": "#/components/schemas/ExerciseSet"
          }
        },
        "additionalProperties": false
      },
      "GetExerciseUnitListItemQuery": {
        "type": "object",
        "properties": {
          "sessionDate": {
            "type": "string",
            "format": "date-time"
          },
          "exerciseSet": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseSet"
            },
            "nullable": true
          },
          "extraSet": {
            "$ref": "#/components/schemas/ExerciseSet"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Wprowadź token JWT w formacie: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}